% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHfitting.R
\name{BHfitting}
\alias{BHfitting}
\title{BHfitting function}
\usage{
BHfitting(
  BH.code,
  dd,
  K.prior,
  r0.prior,
  d.prior,
  N0.prior,
  sdev.prior,
  cores.to.use,
  iter,
  warmup,
  chains
)
}
\arguments{
\item{BH.code}{Stan code generated by the GenerateBHcode functions}

\item{dd}{Dataframe with columns for population size data (colname=popsize), time data (colname=time) and unique identifiers for each population (colname=ident)}

\item{K.prior}{Prior value for the mean carrying capacity (K). Must be on log scale and numeric}

\item{r0.prior}{Prior value for the mean intrinsic rate of growth (r0). Must be on log scale and numeric}

\item{d.prior}{Prior value for the mean death rate (d). Must be on log scale and numeric}

\item{N0.prior}{Prior value for the mean starting population size (N0). Must be on log scale and numeric}

\item{sdev.prior}{Prior value for the standard deviation for the model fitting. Must be numeric. Defaults to 1.}

\item{cores.to.use}{Available cores to do the bayesian models}

\item{iter}{Number of iterations to run for the model, defaults to 1e4. Must be integer}

\item{warmup}{Number of iterations to run warmup. Defaults to 1e3. Must be integer and smaller than warmup}

\item{chains}{Number of chains to run for each fit. Must be integer. Defaults to 1}

\item{Ksd.prior}{Prior value for the standard deviation on carrying capacity (K). Must be on log scale and numeric}

\item{r0sd.prior}{Prior value for the standard deviation on intrinsic rate of growth (r0). Must be on log scale and numeric}

\item{N0sd.prior}{Prior value for the standard deviation on starting population size (N0). Must be on log scale and numeric}
}
\description{
Function to do the actual parallelised Beverton holt fitting of the models
}
\examples{
BHfitting()
}
\keyword{Bayesian}
\keyword{Beverton}
\keyword{Holt}
\keyword{fitting}
\keyword{model}
\keyword{parallel}
