% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratBHcode.R
\name{GenerateBHcode}
\alias{GenerateBHcode}
\title{GenerateBHcode function}
\usage{
GenerateBHcode(K.prior, Ksd.prior, r0.prior, r0sd.prior, d.prior,
  dsd.prior, N0.prior, N0sd.prior, sdev.prior)
}
\arguments{
\item{K.prior}{Prior value for the mean carrying capacity (K). Must be on log scale and numeric}

\item{Ksd.prior}{Prior value for the standard deviation on carrying capacity (K). Must be on log scale and numeric}

\item{r0.prior}{Prior value for the mean intrinsic rate of growth (r0). Must be on log scale and numeric}

\item{r0sd.prior}{Prior value for the standard deviation on intrinsic rate of growth (r0). Must be on log scale and numeric}

\item{d.prior}{Prior value for the mean death rate (d). Must be on log scale and numeric}

\item{N0.prior}{Prior value for the mean starting population size (N0). Must be on log scale and numeric}

\item{N0sd.prior}{Prior value for the standard deviation on starting population size (N0). Must be on log scale and numeric}

\item{sdev.prior}{Prior value for the standard deviation for the model fitting. Must be numeric. Defaults to 1.}

\item{dd}{Dataframe with columns for population size data (colname=popsize), time data (colname=time) and unique identifiers for each population (colname=ident)}

\item{r0sd.prior}{Prior value for the standard deviation on death rate (d). Must be on log scale and numeric}
}
\description{
Generates the stan code to be used for the models, basd on the prior information provided. This allows to only compile one model for all data, saving a lot of computing time.
}
\examples{
GenerateBHcode()
}
\keyword{Beverton}
\keyword{Holt}
\keyword{Stan}
\keyword{code}
