% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCores.R
\name{GetCores}
\alias{GetCores}
\title{GetCores function}
\usage{
GetCores(cores)
}
\arguments{
\item{cores}{Optional argument detailing the number of cores to use. Must be integer. Defaults to NA, in which case all available cores - 1 will be used}
}
\description{
Calculates the available cores based on the user input and machine available cores
}
\examples{
GetCores()
}
\keyword{Beverton}
\keyword{Holt}
\keyword{Stan}
\keyword{code}
