% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPackages.R
\name{loadPackages}
\alias{loadPackages}
\title{LoadPackage function}
\usage{
loadPackages()
}
\description{
Performs all steps in order to fit the BeretonHolt population growth model, show the figures, and return the asked for output
}
\examples{
LoadPackages()
}
\keyword{Bayesian}
\keyword{Beverton}
\keyword{Holt}
\keyword{libraries}
\keyword{model}
\keyword{parallel}
