% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFits.R
\name{PlotFits}
\alias{PlotFits}
\title{PlotFits function}
\usage{
PlotFits(fulloutput, dd, filename)
}
\arguments{
\item{fulloutput}{Output generated in the fitting function, containing all the posterior data from the model fitting}

\item{dd}{Dataframe with columns for population size data (colname=popsize), time data (colname=time) and unique identifiers for each population (colname=ident)}

\item{graphname}{Path with filename to save the figure showing the fits of the data with the posterior predictions. Defaults to NA, in which the plot will not be saved}
}
\description{
Plots the posterior predictions on the actual data so model fit can be checked visually
}
\examples{
PlotFits()
}
\keyword{Population}
\keyword{fit}
\keyword{growth}
\keyword{model}
\keyword{plot}
