% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rK.R
\name{rK}
\alias{rK}
\title{r-K function}
\usage{
rK(
  dd,
  K.prior,
  Ksd.prior,
  r0.prior,
  r0sd.prior,
  N0.prior,
  N0sd.prior,
  sdev.prior = 1,
  cores = NA,
  iter = 10000,
  warmup = 1000,
  chains = 1,
  graphname = NA,
  outputtype = "summary"
)
}
\arguments{
\item{dd}{Dataframe with columns for population size data (colname=popsize), time data (colname=time) and unique identifiers for each population (colname=ident)}

\item{K.prior}{Prior value for the mean carrying capacity (K). Must be on log scale and numeric}

\item{Ksd.prior}{Prior value for the standard deviation on carrying capacity (K). Must be on log scale and numeric}

\item{r0.prior}{Prior value for the mean intrinsic rate of growth (r0). Must be on log scale and numeric}

\item{r0sd.prior}{Prior value for the standard deviation on intrinsic rate of growth (r0). Must be on log scale and numeric}

\item{N0.prior}{Prior value for the mean starting population size (N0). Must be on log scale and numeric}

\item{N0sd.prior}{Prior value for the standard deviation on starting population size (N0). Must be on log scale and numeric}

\item{sdev.prior}{Prior value for the standard deviation for the model fitting. Must be numeric. Defaults to 1.}

\item{cores}{Optional argument detailing the number of cores to use. Must be integer. Defaults to NA, in which case all available cores - 1 will be used}

\item{iter}{Number of iterations to run for the model, defaults to 1e4. Must be integer}

\item{warmup}{Number of iterations to run warmup. Defaults to 1e3. Must be integer and smaller than warmup}

\item{chains}{Number of chains to run for each fit. Must be integer. Defaults to 1}

\item{graphname}{Path with filename to save the figure showing the fits of the data with the posterior predictions. Defaults to NA, in which the plot will not be saved}

\item{outputtype}{Type of output that will be returned. Must be "summary", "full" or "both". summary return the summarised posteriors (mean and sd on log scale), full returns dataframe containing lists of all the posterior samples, both returns a list with both the summary and full output. Defaults to summary}
}
\description{
Performs all steps in order to fit the BeretonHolt population growth model, show the figures, and return the asked for output
}
\examples{
rK()
}
\keyword{Bayesian}
\keyword{fitting}
\keyword{model}
\keyword{parallel}
\keyword{r-K}
