% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizePosteriors.R
\name{SummarizePosteriors}
\alias{SummarizePosteriors}
\title{SumarizePosteriors function}
\usage{
SummarizePosteriors(fulloutput)
}
\arguments{
\item{fulloutput}{Output generated in the BHfitting function, containing all the posterior data from the model fitting}
}
\description{
Function to summarize the posterior data
}
\examples{
SummarizePosteriors()
}
\keyword{Beverton}
\keyword{Holt}
\keyword{output}
\keyword{posteriors}
\keyword{summary}
